+++ libromi/test/mock/mock_joystick.h
@@ -5,8 +5,8 @@ class MockJoystick : public romi::Joystick
 {
 public:
         MOCK_METHOD(romi::JoystickEvent&, get_next_event, (), (override));
-        MOCK_METHOD(int, get_num_axes, (), (override));
-        MOCK_METHOD(double, get_axis, (int i), (override));
-        MOCK_METHOD(int, get_num_buttons, (), (override));
-        MOCK_METHOD(bool, is_button_pressed, (int i), (override));
+        MOCK_METHOD(size_t, get_num_axes, (), (override));
+        MOCK_METHOD(double, get_axis, (size_t i), (override));
+        MOCK_METHOD(size_t, get_num_buttons, (), (override));
+        MOCK_METHOD(bool, is_button_pressed, (size_t i), (override));
 };
